+++ apps/knockdog/src/features/compare/index.ts
@@ -16,5 +16,7 @@ export { DistanceDetailedItem } from './ui/DistanceDetailedItem';
 
 export { PickdropSection } from './ui/PickdropSection';
 
+export { HolidaySection } from './ui/HolidaySection';
+
 export { ComparisonSimpleItem } from './ui/ComparisonSimpleItem';
 export { ComparisonDaysItem } from './ui/ComparisonDaysItem';
